feat: create transaction model

+++ src/database/migrations/1604140110698-CreateTransactions.ts
@@ -26,7 +26,7 @@ export default class CreateTransactions1604140110698
           },
           {
             name: 'value',
-            type: 'float',
+            type: 'decimal',
           },
           {
             name: 'type',
